splits the type coercion from schema module
