+++ dlt/common/data_types/typing.py
@@ -1,0 +1,5 @@
+from typing import Literal, Set, get_args
+
+
+TDataType = Literal["text", "double", "bool", "timestamp", "bigint", "binary", "complex", "decimal", "wei"]
+DATA_TYPES: Set[TDataType] = set(get_args(TDataType))
